Ajouté des composants

+++ src/app/member/member.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { MemberService } from '../../services/member.service';
 import { Member } from '../../models/Member';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmDialogComponent } from '../confirm-dialog/confirm-dialog.component';
 @Component({
   selector: 'app-member',
   templateUrl: './member.component.html',
@@ -17,7 +19,10 @@ export class MemberComponent implements OnInit {
     'action',
   ];
   // Injection de MemberService dans le composant
-  constructor(private memberService: MemberService) {}
+  constructor(
+    private memberService: MemberService,
+    private dialog: MatDialog
+  ) {}
 
   ngOnInit() {
     this.memberService.getAllMembers().subscribe((data) => {
@@ -25,8 +30,15 @@ export class MemberComponent implements OnInit {
     });
   }
   deleteMember(id: string) {
-    this.memberService.deleteMember(id).subscribe(() => {
-      this.dataSource = this.dataSource.filter((member) => member.id !== id);
+    let dialogRef = this.dialog.open(ConfirmDialogComponent);
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.memberService.deleteMember(id).subscribe(() => {
+          this.dataSource = this.dataSource.filter(
+            (member) => member.id !== id
+          );
+        });
+      }
     });
   }
 }
